chore: add profile end point
