+++ app/javascript/dashboard/routes/dashboard/settings/profile/AccessToken.vue
