+++ app/javascript/dashboard/routes/dashboard/settings/profile/ChangePassword.vue
@@ -1,82 +1,73 @@
 <template>
   <form @submit.prevent="changePassword()">
-    <div
-      class="profile--settings--row text-black-900 dark:text-slate-300 flex items-center"
-    >
-      <div class="w-1/4">
-        <h4 class="text-lg text-black-900 dark:text-slate-200">
-          {{ $t('PROFILE_SETTINGS.FORM.PASSWORD_SECTION.TITLE') }}
-        </h4>
-        <p>{{ $t('PROFILE_SETTINGS.FORM.PASSWORD_SECTION.NOTE') }}</p>
-      </div>
-      <div class="w-[45%] p-4">
-        <woot-input
-          v-model="currentPassword"
-          type="password"
-          :class="{ error: $v.currentPassword.$error }"
-          :label="$t('PROFILE_SETTINGS.FORM.CURRENT_PASSWORD.LABEL')"
-          :placeholder="
-            $t('PROFILE_SETTINGS.FORM.CURRENT_PASSWORD.PLACEHOLDER')
-          "
-          :error="
-            $v.currentPassword.$error
-              ? $t('PROFILE_SETTINGS.FORM.CURRENT_PASSWORD.ERROR')
-              : ''
-          "
-          @blur="$v.currentPassword.$touch"
-        />
+    <div class="flex flex-col w-full gap-4">
+      <woot-input
+        v-model="currentPassword"
+        type="password"
+        :styles="inputStyles"
+        :class="{ error: $v.currentPassword.$error }"
+        :label="$t('PROFILE_SETTINGS.FORM.CURRENT_PASSWORD.LABEL')"
+        :placeholder="$t('PROFILE_SETTINGS.FORM.CURRENT_PASSWORD.PLACEHOLDER')"
+        :error="`${
+          $v.currentPassword.$error
+            ? $t('PROFILE_SETTINGS.FORM.CURRENT_PASSWORD.ERROR')
+            : ''
+        }`"
+        @input="$v.currentPassword.$touch"
+      />
 
-        <woot-input
-          v-model="password"
-          type="password"
-          :class="{ error: $v.password.$error }"
-          :label="$t('PROFILE_SETTINGS.FORM.PASSWORD.LABEL')"
-          :placeholder="$t('PROFILE_SETTINGS.FORM.PASSWORD.PLACEHOLDER')"
-          :error="
-            $v.password.$error ? $t('PROFILE_SETTINGS.FORM.PASSWORD.ERROR') : ''
-          "
-          @blur="$v.password.$touch"
-        />
+      <woot-input
+        v-model="password"
+        type="password"
+        :styles="inputStyles"
+        :class="{ error: $v.password.$error }"
+        :label="$t('PROFILE_SETTINGS.FORM.PASSWORD.LABEL')"
+        :placeholder="$t('PROFILE_SETTINGS.FORM.PASSWORD.PLACEHOLDER')"
+        :error="`${
+          $v.password.$error ? $t('PROFILE_SETTINGS.FORM.PASSWORD.ERROR') : ''
+        }`"
+        @input="$v.password.$touch"
+      />
 
-        <woot-input
-          v-model="passwordConfirmation"
-          type="password"
-          :class="{ error: $v.passwordConfirmation.$error }"
-          :label="$t('PROFILE_SETTINGS.FORM.PASSWORD_CONFIRMATION.LABEL')"
-          :placeholder="
-            $t('PROFILE_SETTINGS.FORM.PASSWORD_CONFIRMATION.PLACEHOLDER')
-          "
-          :error="
-            $v.passwordConfirmation.$error
-              ? $t('PROFILE_SETTINGS.FORM.PASSWORD_CONFIRMATION.ERROR')
-              : ''
-          "
-          @blur="$v.passwordConfirmation.$touch"
-        />
+      <woot-input
+        v-model="passwordConfirmation"
+        type="password"
+        :styles="inputStyles"
+        :class="{ error: $v.passwordConfirmation.$error }"
+        :label="$t('PROFILE_SETTINGS.FORM.PASSWORD_CONFIRMATION.LABEL')"
+        :placeholder="
+          $t('PROFILE_SETTINGS.FORM.PASSWORD_CONFIRMATION.PLACEHOLDER')
+        "
+        :error="`${
+          $v.passwordConfirmation.$error
+            ? $t('PROFILE_SETTINGS.FORM.PASSWORD_CONFIRMATION.ERROR')
+            : ''
+        }`"
+        @input="$v.passwordConfirmation.$touch"
+      />
 
-        <woot-button
-          :is-loading="isPasswordChanging"
-          type="submit"
-          :disabled="
-            !currentPassword ||
-            !passwordConfirmation ||
-            !$v.passwordConfirmation.isEqPassword
-          "
-        >
-          {{ $t('PROFILE_SETTINGS.FORM.PASSWORD_SECTION.BTN_TEXT') }}
-        </woot-button>
-      </div>
+      <form-button
+        type="submit"
+        color-scheme="primary"
+        variant="solid"
+        size="large"
+        :disabled="isButtonDisabled"
+      >
+        {{ $t('PROFILE_SETTINGS.FORM.PASSWORD_SECTION.BTN_TEXT') }}
+      </form-button>
     </div>
   </form>
 </template>
 
 <script>
 import { required, minLength } from 'vuelidate/lib/validators';
-import { mapGetters } from 'vuex';
 import alertMixin from 'shared/mixins/alertMixin';
 import { parseAPIErrorResponse } from 'dashboard/store/utils/api';
-
+import FormButton from 'v3/components/Form/Button.vue';
 export default {
+  components: {
+    FormButton,
+  },
   mixins: [alertMixin],
   data() {
     return {
@@ -85,6 +76,12 @@ export default {
       passwordConfirmation: '',
       isPasswordChanging: false,
       errorMessage: '',
+      inputStyles: {
+        borderRadius: '12px',
+        padding: '6px 12px',
+        fontSize: '14px',
+        marginBottom: '2px',
+      },
     };
   },
   validations: {
@@ -105,10 +102,13 @@ export default {
     },
   },
   computed: {
-    ...mapGetters({
-      currentUser: 'getCurrentUser',
-      currentUserId: 'getCurrentUserID',
-    }),
+    isButtonDisabled() {
+      return (
+        !this.currentPassword ||
+        !this.passwordConfirmation ||
+        !this.$v.passwordConfirmation.isEqPassword
+      );
+    },
   },
   methods: {
     async changePassword() {
@@ -117,38 +117,21 @@ export default {
         this.showAlert(this.$t('PROFILE_SETTINGS.FORM.ERROR'));
         return;
       }
-
+      let alertMessage = this.$t('PROFILE_SETTINGS.PASSWORD_UPDATE_SUCCESS');
       try {
         await this.$store.dispatch('updateProfile', {
           password: this.password,
           password_confirmation: this.passwordConfirmation,
           current_password: this.currentPassword,
         });
-        this.errorMessage = this.$t('PROFILE_SETTINGS.PASSWORD_UPDATE_SUCCESS');
       } catch (error) {
-        this.errorMessage =
+        alertMessage =
           parseAPIErrorResponse(error) ||
           this.$t('RESET_PASSWORD.API.ERROR_MESSAGE');
       } finally {
-        this.isPasswordChanging = false;
-        this.showAlert(this.errorMessage);
+        this.showAlert(alertMessage);
       }
     },
   },
 };
 </script>
-
-<style lang="scss">
-@import '~dashboard/assets/scss/mixins.scss';
-
-.profile--settings--row {
-  @include border-normal-bottom;
-  padding: var(--space-normal);
-  .small-3 {
-    padding: var(--space-normal) var(--space-medium) var(--space-normal) 0;
-  }
-  .small-9 {
-    padding: var(--space-normal);
-  }
-}
-</style>
