+++ app/javascript/dashboard/routes/dashboard/settings/profile/HotKeyCard.vue
