+++ app/javascript/dashboard/routes/dashboard/settings/profile/NotificationCheckBox.vue
