+++ app/javascript/dashboard/routes/dashboard/settings/profile/NotificationPreferences.vue
