+++ app/javascript/dashboard/routes/dashboard/settings/profile/UserProfilePicture.vue
