+++ app/javascript/dashboard/routes/dashboard/settings/profile/constants.js
